refactor(device-transfer): migrate DeviceTransferValidator to DeviceTransferGuard pattern
Migrate DeviceTransferValidator to DeviceTransferGuard applying the guard clause pattern
to reduce ambiguity.

- Renamed DeviceTransferValidator to DeviceTransferGuard
- Updated imports and usages in AcceptDeviceTransferAction,
CancelDeviceTransferAction, CreateDeviceTransferAction and
RejectDeviceTransferAction
