+++ app/Actions/DeviceTransfer/Cancel/CancelDeviceTransferAction.php
@@ -2,10 +2,10 @@
 
 namespace App\Actions\DeviceTransfer\Cancel;
 
-use App\Actions\Validator\DeviceTransferValidator;
 use App\Enums\Device\DeviceTransferStatus;
 use App\Exceptions\Application\DeviceTransfer\CancelDeviceTransferFailedException;
 use App\Exceptions\BusinessRules\BusinessRuleException;
+use App\Guards\DeviceTransferGuard;
 use App\Models\DeviceTransfer;
 use App\Models\User;
 use Illuminate\Support\Facades\DB;
@@ -17,7 +17,7 @@ class CancelDeviceTransferAction
     public function __invoke(User $user, DeviceTransfer $transfer): DeviceTransfer
     {
         try {
-            $this->validateBusinessRules($user, $transfer);
+            $this->enforceBusinessRules($user, $transfer);
 
             return DB::transaction(function () use ($user, $transfer) {
                 $this->cancelTransfer($transfer);
@@ -32,10 +32,10 @@ class CancelDeviceTransferAction
         }
     }
 
-    private function validateBusinessRules(User $user, DeviceTransfer $transfer): void
+    private function enforceBusinessRules(User $user, DeviceTransfer $transfer): void
     {
-        DeviceTransferValidator::mustBeSender($user, $transfer);
-        DeviceTransferValidator::mustBePending($transfer);
+        DeviceTransferGuard::mustBeSender($user, $transfer);
+        DeviceTransferGuard::mustBePending($transfer);
     }
 
     private function cancelTransfer(DeviceTransfer $transfer): void
